Turn on warnings as errors (for windows), fix a few warnings

+++ include/mega/sync.h
@@ -750,8 +750,8 @@ struct Syncs
     UnifiedSync* appendNewSync(const SyncConfig&, MegaClient& mc);
 
     bool hasRunningSyncs();
-    unsigned numRunningSyncs();
-    unsigned numSyncs();    // includes non-running syncs, but configured
+    size_t numRunningSyncs();
+    size_t numSyncs();    // includes non-running syncs, but configured
     Sync* firstRunningSync();
     Sync* runningSyncByBackupId(handle backupId) const;
     SyncConfig* syncConfigByBackupId(handle backupId) const;
